// FE/src/api/models/createPractitionerDto.ts
/**
 * Generated by orval v7.13.1 🍺
 * Do not edit manually.
 * SIGREF.API
 * OpenAPI spec version: 1.0
 */
import type { IdentifierDto } from "./identifierDto";
import type { HumanNameDto } from "./humanNameDto";
import type { ContactPointDto } from "./contactPointDto";
import type { AdministrativeGender } from "./administrativeGender";

export interface CreatePractitionerDto {
  /** @nullable */
  identifier?: IdentifierDto[] | null;
  active: boolean;
  name: HumanNameDto[];
  /** @nullable */
  telecom?: ContactPointDto[] | null;
  gender?: AdministrativeGender;
  /** @nullable */
  birthDate?: string | null;
}