// FE/src/api/models/updatePractitionerDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SIGREF.API
 * OpenAPI spec version: 1.0
 */
import type { IdentifierDto } from "./identifierDto";
import type { HumanNameDto } from "./humanNameDto";
import type { ContactPointDto } from "./contactPointDto";
import type { AdministrativeGender } from "./administrativeGender";

export interface UpdatePractitionerDto {
  /** @nullable */
  identifier?: IdentifierDto[] | null;
  active: boolean;
  name: HumanNameDto[];
  /** @nullable */
  telecom?: ContactPointDto[] | null;
  gender?: AdministrativeGender;
  /** @nullable */
  birthDate?: string | null;
}